chore: Add new lyrics to _messages.json
New lyrics have been added to the _messages.json file. This change includes the addition of the line "All within my hands, squeeze it in, crush it down." to the list of lyrics. This update enhances the content of the page and provides a more complete experience for the users.